requirements.txt
Brotli 
certifi 
cffi 
charset-normalizer 
ConfigArgParse==1.7
h2 
hpack==4.0.0
hyperframe 
idna 
imageio
imageio-ffmpeg==0.4.7
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
libigl==2.5.1
lpips==0.1.4
markdown-it-py==3.0.0
mdurl==0.1.2
networkx==3.4.1
ninja==1.11.1.1
numpy 
opencv-python==4.10.0.84
packaging==24.1
Pillow 
plyfile==1.1
pybind11==2.13.6
pycparser 
Pygments==2.18.0
PySocks 
requests 
rich==13.9.2
scikit-image==0.24.0
scipy==1.14.1
tifffile==2024.9.20
tqdm==4.66.5
typing_extensions 
urllib3 
zstandard==0.23.0